Updating answer (in Python)
I managed to get through the first three sections of this problem, but had trouble on the 4th section (Identifying the frequencies of codons). I looked throughout the internet for a solution and came across this one: 
https://stackoverflow.com/questions/46385576/python-how-to-count-frequency-of-triple-nucleotide

I couldn't get it to work, but I still left my incomplete code.